calculated bootstrapping (1000x)
