+++ .idea/misc.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.6.3 (C:\Users\chris\Miniconda3\envs\slovenia_env\python.exe)" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.6.3 (C:\Users\Admin\Miniconda3\envs\slovenia_env\python.exe)" project-jdk-type="Python SDK" />
 </project>
